Fix compile error on windows
static_cast an int64 to int to suppress warning. The value is known to be <17

TBR=

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@75046 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 7ba168d309dd2777cd07bbbd3c024c26265f79ba

+++ chrome/renderer/render_widget.cc
@@ -510,8 +510,8 @@ void RenderWidget::AnimateIfNeeded() {
       // avoid exposing this delay to javascript, we keep posting delayed
       // tasks until base::Time::Now() has advanced far enough.
       int64 delay = (animation_floor_time_ - now).InMillisecondsRoundedUp();
-      UMA_HISTOGRAM_CUSTOM_COUNTS("Renderer4.AnimationDelayTime", delay,
-                                  0, 16, 17);
+      UMA_HISTOGRAM_CUSTOM_COUNTS("Renderer4.AnimationDelayTime",
+                                  static_cast<int>(delay), 0, 16, 17);
       animation_task_posted_ = true;
       MessageLoop::current()->PostDelayedTask(FROM_HERE, NewRunnableMethod(
           this, &RenderWidget::AnimationCallback), delay);
